Refactor who we are grid layout
